feat(web-server)!: move everything over to /api
This is still in the move fast and break a lot phase, hence the gigantic
diff. I'll get better about it, I promise. This commit shifts to using
the `/api` endpoint. This is so that I can have a seperate web server
purely for serving up and doing front-end things without worrying about
the actual logic of what the hell is going on.

Long-term, this is going to be running on `api.<domain>.com`, and not
localhost, but it's fine for now.